New neteyeshare installation path and secure agent deploy.

+++ scripts/011_init_neteyeshare_weblink.sh
@@ -1,33 +1,41 @@
 #!/bin/bash
 
-HTTP_PASSWD_FOLDER="/usr/local/httpd"
-HTTP_PASSWD_FILE="${HTTP_PASSWD_FOLDER}/passwd"
+HTTP_PASSWD_FOLDER="/neteye/shared/httpd"
+HTTP_PASSWD_FILE="${HTTP_PASSWD_FOLDER}/.htpasswd"
 HTTP_CONF_FILE="/etc/httpd/conf.d/neteye-share.conf"
+PWD_SHARE_LOGIN="/root/.pwd_neteye_configro"
 
-HTTP_USERNAME="sharero"
-HTTP_PASSWORD="R3ad0nLy_#12"
+HTTP_USERNAME="configro"
+#HTTP_PASSWORD="R3ad0nLy_#12"
+HTTP_PASSWORD=`head /dev/urandom | tr -dc A-Za-z0-9 | head -c 13 ; echo ''`
+
 
 #Create folder structure
-if [ ! -d $HTTP_PASSWD_FOLDER ]
+if [ ! -f ${HTTP_PASSWD_FILE} ]
 then
 
-   echo "[+] Initializing NetEye-Share: Web Alias"
-   echo "    Creating Folder ${HTTP_PASSWD_FOLDER}"
-   mkdir ${HTTP_PASSWD_FOLDER}
+   echo "[+] 011: Configuration of neteyeshare/: registragion of httpd alias and creation of new HTTP user: ${HTTP_USERNAME}."
+   echo "    Setup of authentication for user: ${HTTP_USERNAME} in: ${HTTP_PASSWD_FILE}"
+   mkdir -p ${HTTP_PASSWD_FOLDER}
 
-   echo "[+] Registering new HTTP user: ${HTTP_USERNAME}"
+   # Create new httpd user with password to protect some folders
    htpasswd -b  -c ${HTTP_PASSWD_FILE} ${HTTP_USERNAME} ${HTTP_PASSWORD}
-fi
 
-if [ ! -f ${HTTP_CONF_FILE} ]
-then
-   cp neteye4/neteyeshare/neteye-share.conf  ${HTTP_CONF_FILE}
+   if [ ! -f ${HTTP_CONF_FILE} ]
+   then
+      cp neteye4/neteyeshare/neteye-share.conf  ${HTTP_CONF_FILE}
+   fi
+
+   echo ${HTTP_PASSWORD} >> ${PWD_SHARE_LOGIN} 
 
    echo "[!] Now please reload service httpd to activate new neteyeshare weblink"
-   echo "    systemctl restart httpd.service"
+   echo "    Hint: systemctl restart httpd.service"
    echo "  "
-   echo "[i] Login on web via: https://neteye_fqdn/neteyeshare"
+   echo "[i] The neteyeshare comes with no authentication and can be accessed on web via: https://neteye_fqdn/neteyeshare"
+   echo "    Some configuration folders are protected by these credentials: (i.e. ./monitoring/agents/microsoft/icinga/configs/)"
    echo "    Username: ${HTTP_USERNAME}"
    echo "    Password: ${HTTP_PASSWORD}"
+   echo "  "
+   echo "    The password is stored in: ${PWD_SHARE_LOGIN}"
    echo "  "
 fi
